* Fix space between bottom and ChatContainer * Set max idle timeout to infinity (maybe after we should implement health check to connection to prevent zombie connections) * Fix VAD model path * Added toast notifications when add accounts * Fix log out function * Added delivery ack

+++ src/api/device/device.rs
@@ -427,7 +427,7 @@ impl Device {
     /// Generate a random message identifier
     pub fn generate_message_id() -> u64 {
         let mut rng = rand::rng();
-        rng.random_range(0..u64::MAX)
+        rng.random_range(0..i64::MAX as u64)
     }
 }
